factor out mem-model-only util
